+++ src/motion/functionality/types.ts
@@ -1,10 +1,10 @@
 import { RefObject, ComponentType, CSSProperties, ReactElement } from "react"
 import { MotionProps } from "../types"
-import { ComponentAnimationControls } from "../../animation/ComponentAnimationControls"
+import { ValueAnimationControls } from "../../animation/ValueAnimationControls"
 import { MotionValuesMap } from "../utils/use-motion-values"
 
 export interface FunctionalProps extends MotionProps {
-    controls: ComponentAnimationControls
+    controls: ValueAnimationControls
     values: MotionValuesMap
     innerRef: RefObject<Element | null>
 }
@@ -17,7 +17,7 @@ export interface FunctionalComponentDefinition {
 export type UseFunctionalityComponents<P = {}> = (
     props: P & MotionProps,
     values: MotionValuesMap,
-    controls: ComponentAnimationControls<P>,
+    controls: ValueAnimationControls<P>,
     ref: RefObject<Element | null>,
     style: CSSProperties,
     isStatic?: boolean