+++ CHANGELOG.md
@@ -25,7 +25,7 @@ Framer Motion adheres to [Semantic Versioning](http://semver.org/).
 
 -   Using `transition` and `transitionEnd` as direct values on the `animate` prop on subsequent renders.
 -   Rounding `zIndex`.
--   `AnimationControls.start` now accepts the same `AnimationDefinition` as `ComponentAnimationControls.start`.
+-   `AnimationControls.start` now accepts the same `AnimationDefinition` as `ValueAnimationControls.start`.
 
 ## [0.16.8] 2019-05-06